+++ app/views/events/create.blade.php
@@ -34,6 +34,6 @@ Create Event
 
 @section('js')
 {{ Asset::scripts('form') }}
-{{ HTML::script('//cdnjs.cloudflare.com/ajax/libs/moment.js/2.5.1/moment.min.js') }}
+{{ HTML::script('//cdnjs.cloudflare.com/ajax/libs/moment.js/2.6.0/moment.min.js') }}
 {{ Asset::scripts('picker') }}
 @stop