+++ app/views/posts/index.blade.php
@@ -21,13 +21,13 @@ Blog
             @endif
         </p>
     </div>
-    @if (Credentials::check() && Credentials::hasAccess('blog'))
+    @auth('blog')
         <div class="col-xs-4">
             <div class="pull-right">
                 <a class="btn btn-primary" href="{{ URL::route('blog.posts.create') }}"><i class="fa fa-book"></i> New Post</a>
             </div>
         </div>
-    @endif
+    @endauth
 </div>
 @foreach($posts as $post)
     <h2>{{ $post->title }}</h2>
@@ -36,9 +36,9 @@ Blog
     </p>
     <p>
         <a class="btn btn-success" href="{{ URL::route('blog.posts.show', array('posts' => $post->id)) }}"><i class="fa fa-file-text"></i> Show Post</a>
-        @if (Credentials::check() && Credentials::hasAccess('blog'))
+        @auth('blog')
              <a class="btn btn-info" href="{{ URL::route('blog.posts.edit', array('posts' => $post->id)) }}"><i class="fa fa-pencil-square-o"></i> Edit Post</a> <a class="btn btn-danger" href="#delete_post_{{ $post->id }}" data-toggle="modal" data-target="#delete_post_{{ $post->id }}"><i class="fa fa-times"></i> Delete Post</a>
-        @endif
+        @endauth
     </p>
     <br>
 @endforeach
@@ -46,7 +46,7 @@ Blog
 @stop
 
 @section('bottom')
-@if (Credentials::check() && Credentials::hasAccess('blog'))
+@auth('blog')
     @include('posts.deletes')
-@endif
+@endauth
 @stop